+++ packages/proto-gen-dts/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.2.0](https://github.com/planjs/stan/compare/proto-gen-dts@0.1.3...proto-gen-dts@0.2.0) (2021-02-03)
+
+
+### Bug Fixes
+
+* **proto-gen-ts:** reduce repeated parsing of proto ([0b80264](https://github.com/planjs/stan/commit/0b80264338a09aecc0dcc52a598147abbf336fce))
+
+
+### Features
+
+* **proto-gen-dts:** add generate referenceEntryFile ([f650a44](https://github.com/planjs/stan/commit/f650a4415059e558ccc1affe9a495ec64408cf03))
+
+
+
+
+
 ## [0.1.3](https://github.com/planjs/stan/compare/proto-gen-dts@0.1.2...proto-gen-dts@0.1.3) (2021-02-03)